pre-token-changes
