+++ src/main.rs
@@ -1,7 +1,8 @@
-mod lexer;
 mod cli;
 mod err;
 mod jace_file;
+mod lexer;
+mod parser;
 
 use std::fs::File;
 use std::io::Read;
@@ -11,8 +12,7 @@ use clap::Parser;
 use cli::Cli;
 use cli::Command;
 use jace_file::JaceFile;
-use lexer::Lexer;
-use lexer::LexerIterator;
+use crate::lexer::prelude::*;
 
 fn main() {
     //let args = Cli::parse();
@@ -24,28 +24,60 @@ fn main() {
 
     let jcf = JaceFile::new("test.jc", 
         r#"
+            def map :: (a => b), [a] => [b]
+            where
+                a : Number
+                b : Number
+            in
+            case
+                _, {} => {}
+                f, {x:xs} => f x! : map f xs!
+
+            type Number :: Float | Integer
+
+            def sum :: a, a, a, a => a
+            where
+                a : Number
+            in
+                a, b, c, d => let
+                    inter := a + b
+                    inter2 := c + d
+                in inter + inter2
+
+            def complexOperation :: Number, Number, Number => Number
+                a, b, c => a^2+a*b+2^2*5^2/57*56
+
+            def multi_return :: Number, Number, Number, Number => Number
+                a, b, c, d => a*5+b*7,c*3+d*4
+
+            type Foo ::
+                bar : String
+                baz : Number
+
+            const MATH_PI :: 3.14
+
             class Equal a ::
                 (==) :: a, a => Bool
-                (!=) :: a, a => Bool
-                (>=) :: a, a => Bool
-                (<=) :: a, a => Bool
-                (+)  :: a, a => a
-                (-)  :: a, a => a 
-                (*)  :: a, a => a 
-                (/)  :: a, a => a 
 
-            def doOperation::Integer,Integer=>Integer
-                a,b=>a+b*(2^5)-2.5+3.6,7
+            def main :: ()
+                foo := {bar = "FooBaz", baz = 67}
+                multi1,multi2 := multi_return 1 2 3 4!
+                {m1, m2} := multi_return 5 6 7 8!
+                arr_ret := {multi_return 5 6 7 8!}
+                arr_ret_x_2 := map (a => a * 2) arr_ret!
 
-            def main::()
-                @
-                print "Hello World!"!
-                john := {name="John", age = 21}
-                john.name
+            main
         "#);
 
     let mut lexer = Lexer::new(jcf).into_iter();
+    let toks: Vec<TokenKind> = lexer
+        .filter_map(|t| t.ok())
+        .map(|t| t.0)
+        .collect();
+    println!("{toks:#?}");
 
+    let (_, (result, errs)) = parser::parse(&toks).unwrap();
+    println!("{result:#?}");
 }
 
 fn compiler_pipeline(path: PathBuf) {
